requirements.txt
requests
pandas
beautifulsoup4
tqdm
python-dotenv
anthropic
